% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BristolBathPlaces}
\alias{BristolBathPlaces}
\title{Places in South-West England}
\format{
A `data frame` object.
\describe{
  \item{place}{City or town}
  \item{lat}{Latitude}
  \item{long}{Longitude}
}
}
\usage{
BristolBathPlaces
}
\description{
A `data frame` object containing several prominent towns and cities in South-West England. This data was extracted from OpenStreetMap.
}
\examples{
BristolBathPlaces
}
\references{
\itemize{
\item Overpass API for OpenStreetMap Data \cr
\url{https://wiki.openstreetmap.org/wiki/Overpass_API}
}
}
\keyword{datasets}
