% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BristolBathPolygons}
\alias{BristolBathPolygons}
\title{Boundaries of Postcode Sectors in South-West England as a SpatialPolygons Object}
\format{
A `SpatialPolygons` object.

Each postcode sector has the following attributes:
\describe{
  \item{Name}{UK postcode sector (e.g. `BS1 1`, `BS6 5`)}
  \item{Description}{Empty string value}
}
}
\usage{
BristolBathPolygons
}
\description{
A `SpatialPolygons` object containing boundaries of Postcode Sectors of the Bristol (BS) and Bath (BA) area in South-West England of the United Kingdom.
}
\examples{
plot(BristolBathPolygons)
}
\keyword{datasets}
