% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vertex_attr_aggregate.R
\name{get_vertex_attr_aggregate}
\alias{get_vertex_attr_aggregate}
\title{Compute Aggregate of Vertex Attributes}
\usage{
get_vertex_attr_aggregate(g, attr, groups, func, ...)
}
\arguments{
\item{g}{An `igraph` object}

\item{attr}{Character value, indicating the vertex attribute to be aggregated}

\item{groups}{Vector of character or factor, which indicate the groups to be aggregated}

\item{func}{Aggregation function}

\item{...}{Optional arguments to be passed onto the aggregation function}
}
\value{
A vector or matrix object
}
\description{
Calculate the group by aggregate of a specific vertex attribute
}
\examples{
# Use the postcode sector to derive the area
# Example: BS = Bristol; BA = Bath
library(stringr)
area <- factor(str_extract(vertex_attr(BristolBathGraph, "name"),"^[A-Z]*"),
               labels = c(BS="Bristol", BA="Bath"))

# Calculate the aggregate of vertex attribute
# Returns a vector object of length 2
# The names represent postcode area
# The value represent total population of each postcode area
get_vertex_attr_aggregate(g = BristolBathGraph,
                          groups = area,
                          attr = "population", 
                          func = sum)

# Optional arguments can be passed onto `func` using `...`.
# This will calculate the percentiles of the population of each postcode area
# Returns a matrix object of 2 columns * 4 rows.
# The columns represent postcode areas
# The rows are names as '20\%', '40\%', '60\%' and '80\%'
# You may use `names=FALSE` to suppress row names (see documentation for `?quantile()`)
get_vertex_attr_aggregate(g = BristolBathGraph,
                          groups = area,
                          attr = "population", 
                          func = quantile,
                          probs = c(0.2, 0.4, 0.6, 0.8))

get_vertex_attr_aggregate(g = BristolBathGraph,
                          groups = area,
                          attr = "population", 
                          func = quantile,
                          probs = c(0.2, 0.4, 0.6, 0.8),
                          names = FALSE)
}
\author{
Timothy Wong, \email{timothy.wong@hotmail.co.uk}
}
