% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_communities.R
\name{merge_communities}
\alias{merge_communities}
\title{Merge Communities in a Graph}
\usage{
merge_communities(
  z,
  g,
  m,
  at_level,
  assign_level,
  vertex_attribute,
  vertex_aggregate_func,
  vertex_aggregate_lower_threshold,
  vertex_aggregate_upper_threshold,
  edge_attribute,
  cost_aggregate_func,
  cost_lower_threshold,
  cost_upper_threshold,
  within_level,
  within_zones = NULL,
  penalty = function(x) {     base::log(scales::rescale(-x) + 1)^0.1 },
  cost_aggregate_args = list(),
  vertex_aggregate_args = list()
)
}
\value{
A `data frame` object
}
\description{
Merge smaller zones to create larger ones.
}
\examples{

}
\references{
\itemize{
}
}
\author{
Timothy Wong, \email{timothy.wong@hotmail.co.uk}
}
